// packages/react-cv/src/components/resume-types.ts
export interface Work {
	// e.g. Facebook
	company: string;

	// e.g. Menlo Park, CA
	location?: string;

	// e.g. Social Media Company
	description?: string;

	// e.g. Software Engineer
	position?: string;

	// e.g. http://facebook.example.com
	url?: string;

	// resume.json uses the ISO 8601 date standard e.g. 2014-06-29
	startDate?: string;

	// e.g. 2012-06-29
	endDate?: string;

	// Give an overview of your responsibilities at the company
	summary?: string;

	// Specify multiple accomplishments, e.g. Increased profits by 20% from 2011-2012 through viral advertising
	highlights?: string[];

	// Keywords or skills, e.g. React, JavaScript
	keywords?: string[];
}
// ...
export interface Skill {
	id?: string;
	// e.g. Web Development
	name: string;

	// "e.g. Master
	level?: string;

	// List some keywords pertaining to this skill, e.g. HTML
	keywords?: string[];
	preferred?: string[]; // Preferred skills
}